% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{extract_timestamps}
\alias{extract_timestamps}
\title{extract_timestamps}
\usage{
extract_timestamps(dataframe)
}
\arguments{
\item{dataframe}{is the previously imported file usually saved as format "Timestamps_of_spikes_60.txt"}
}
\description{
Turns the data points of spikes timed into seconds.
Note that this is for a 19254 effective Sampling Rate, ie. actual recorded SR was 19754 but after timestamp deletion the eff.SR is 19254.
It is important also to check that you have the right number of seconds per chunk, created in Julia, in this case 5 seconds per voltage chunk.
}
