% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{find_active_channels}
\alias{find_active_channels}
\title{find_active_channels}
\usage{
find_active_channels(org_data, dodgy_channels)
}
\arguments{
\item{dodgy_channels}{input any noticeably dodgy channels, e.g. that always seem high in frequency regardless of conditions.}
}
\value{
returns df of the channels with mean frequency over 0.5Hz.
}
\description{
sorts the data frame to find all values with a mean Frequency of >0.5Hz.
}
