% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{find_top50_channels}
\alias{find_top50_channels}
\title{find_top50_channels}
\usage{
find_top50_channels(df, age, organoid, condition)
}
\arguments{
\item{df}{is the dataframe to be filtered. In my case, I have one giant df, so I have other factors to filter by.}

\item{age}{is the desired age of the organoid of interested.}

\item{organoid}{is the specific organoid I am interested in}

\item{condition}{is either Control, PTX or Propofol i.e. drug condition.}

\item{C_temp}{The temperature in degrees Celsius}
}
\value{
Has filtered the original df to only keep the top 50 most active channels.
}
\description{
Ranks the channels in a df by frequency, and returns the top 50 in a new df.
}
\examples{
org2_p60_top50_control <- find_top50_channels(giant.df, "P60", "2", "Control")
}
