% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{make_raster_plot}
\alias{make_raster_plot}
\title{make_raster_plot}
\usage{
make_raster_plot(df, title, Condition)
}
\arguments{
\item{df}{is of course the df with the timestamps you want to plot. This is returned from the extract_timestamps function.}

\item{title}{is the title you would like to appear on the plot "Title1"}

\item{Condition}{is the condition of recording e.g. Control, PTX, Propofol. This also corresponds to the colours of the plots.}
}
\description{
Uses the timestamps to make a rasterplot of the spikes.
}
