% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{plot_the_difference}
\alias{plot_the_difference}
\title{plot_the_difference}
\usage{
plot_the_difference(df.new)
}
\arguments{
\item{df.new}{is the df after subtract_frequency and channel_to_index}
}
\value{
a matrix representing the difference in activity per electrode
}
\description{
Make a heatmap of the difference in activity between two conditions on the HD-MEA.
}
