% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{process_organoid_data}
\alias{process_organoid_data}
\title{process_organoid_data}
\usage{
process_organoid_data(spikes_file, channels_file, condition)
}
\arguments{
\item{spikes_file}{is the .txt file that has the number of spikes measured per channel (row) as a vector.}

\item{channels_file}{is the raw file that has the channel numbers used in this analysis. Only relevant when analysis was performed on a subset of channels.}

\item{condition}{is the condition of the recording ie. in the organoid experiments, Control, PTX or Propofol.}
}
\value{
returns a df with just the top 50 active electrodes
}
\description{
Function does the same as read_org_data but returns just the top 50 most active channels.
}
