% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organoid functions p60.R
\name{read_org_data}
\alias{read_org_data}
\title{read_org_data}
\usage{
read_org_data(spikes_file, channels_file, condition)
}
\arguments{
\item{spikes_file}{is the .txt file that has the number of spikes measured per channel (row) as a vector.}

\item{channels_file}{is the raw file that has the channel numbers used in this analysis. Only relevant when analysis was performed on a subset of channels.}

\item{condition}{is the condition of the recording ie. in the organoid experiments, Control, PTX or Propofol.}
}
\value{
returns useful index of row.column
}
\description{
Function to load organoid spikes data, that I have produced using Julia/Jupyter notebook.
See github.com/chloemhall/HD-MEA-Organoids
}
