% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{subtract_frequency}
\alias{subtract_frequency}
\title{subtract_frequency}
\usage{
subtract_frequency(targetdf, channelsdf)
}
\arguments{
\item{targetdf}{is a df with many channels and you want to search it & find the ones you want, which are in channelsdf}

\item{channelsdf}{is a df that has the channels you want / are interested in}
}
\value{
gives us end_df which is a df where the frequency of x has been subtracted from y.
}
\description{
take channels from the channelsdf and extract those channels from targetdf
}
