% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{channel_to_index}
\alias{channel_to_index}
\title{channel_to_index}
\usage{
channel_to_index(Ch)
}
\arguments{
\item{Ch}{is an integer 1:4096 of the channel / electrode.}
}
\value{
returns useful index of row.column
}
\description{
Converts a simple integer of channel into more useful index
( on 3Brain Accura chips, this is 64 rows x 64 columns. eg. channel 2 is row 1, col 2 ie. 1.2 )
}
\examples{
> channel<- channel_to_index(65)
> print(channel)
> 2 1 #ie. row 2. column 1
}
