% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions .R
\name{subtract_frequency}
\alias{subtract_frequency}
\title{subtract_frequency}
\usage{
subtract_frequency(all.data, age, organoid, condition1, condition2)
}
\arguments{
\item{all.data}{is a df that has all the data in one giant table. You will search and filter it.}

\item{age}{Age of the organoid / experiment you want to select.}

\item{organoid}{The repeat number of the organoid / experiment you want to select.}

\item{condition1}{This is the drug condition you want to subtract from ie. baseline.}

\item{condtion2}{This is the drug condition you want to subtract from condition1}
}
\value{
gives us t3_df which is a df where the frequency of condition2 has been subtracted from condition1.
}
\description{
take channels from the giant dataframe and extracts the channels of interest, performing a subtraction calculation of frequency per channel
}
